[vcr-2.0] Remove unused azure-metrics
There are a lot of metrics not used in VCR today but are still registered and emitted. This exceeds the allowable number of metrics causing exceptions that prevent even the useful ones from being emitted. So just get rid of them but keep them around so as to not break old code.
